Set ChargeLine for each Attendance instance to null when removing Attendance
- If ChargeLine for each Attendance instance is not set to null prior to removing, then Hibernate exception is thrown since ChargeLine and Attendance entities are mapped to each other.

+++ src/main/java/com/aimacademyla/service/impl/CourseSessionServiceImpl.java
@@ -154,13 +154,12 @@ public class CourseSessionServiceImpl extends GenericServiceImpl<CourseSession, 
 
                     attendance.setChargeLine(chargeLine);
                     chargeLineService.addChargeLine(chargeLine);
-                    attendanceDAO.update(attendance);
-                } else {
-                    if (chargeLine != null) {
+                }
+                else if(!attendance.getWasPresent() && chargeLine != null){
                         attendance.setChargeLine(null);
+                        //Update Attendance with null ChargeLine first before removing ChargeLine
                         attendanceDAO.update(attendance);
                         chargeLineService.removeChargeLine(chargeLine);
-                    }
                 }
             }
